V1.1: Menambahkan if else

+++ src/main/java/ws/b/project04/ProductServiceController.java
@@ -9,13 +9,12 @@ import java.util.Map;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
-
-import ws.b.project04.Product;
 
 /**
  *
@@ -39,24 +38,44 @@ public class ProductServiceController {
    
    @RequestMapping(value = "/products/{id}", method = RequestMethod.DELETE)
    public ResponseEntity<Object> delete(@PathVariable("id") String id) { 
-      productRepo.remove(id);
-      return new ResponseEntity<>("Product is deleted successsfully", HttpStatus.OK);
+       if(!productRepo.containsKey(id)){
+           return new ResponseEntity<>("Product deleted not found", HttpStatus.OK);
+       }
+       else{
+           productRepo.remove(id);
+           return new ResponseEntity<>("Product is deleted successsfully", HttpStatus.OK);
+       }
+      
    }
    
    @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
    public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product) { 
-      productRepo.remove(id);
-      product.setId(id);
-      productRepo.put(id, product);
-      return new ResponseEntity<>("Product is updated successsfully", HttpStatus.OK);
+      if(!productRepo.containsKey(id)){
+           return new ResponseEntity<>("Product updated not found", HttpStatus.OK);
+       }
+       else{
+           productRepo.remove(id);
+           product.setId(id);
+           productRepo.put(id, product);
+           return new ResponseEntity<>("Product is updated successsfully", HttpStatus.OK);
+       }
+      
    }
    
    @RequestMapping(value = "/products", method = RequestMethod.POST)
-   public ResponseEntity<Object> createProduct(@RequestBody Product product) {
-      productRepo.put(product.getId(), product);
-      return new ResponseEntity<>("Product is created successfully", HttpStatus.CREATED);
+   public ResponseEntity<Object> createProduct(@RequestBody @Validated Product product) {
+       
+      if(productRepo.containsKey(product.getId())){
+           return new ResponseEntity<>("Product not duplicate", HttpStatus.OK);
+       }
+       else{
+           productRepo.put(product.getId(), product);
+           return new ResponseEntity<>("Product is created successfully", HttpStatus.CREATED);
+       }
+       
+       
+      
    }
-   
    @RequestMapping(value = "/products")
    public ResponseEntity<Object> getProduct() {
       return new ResponseEntity<>(productRepo.values(), HttpStatus.OK);
